+++ packages/linear-bootstrap/src/http-cli.ts
@@ -1,0 +1,3 @@
+import { startHttpServer } from "./http.js";
+
+startHttpServer();